+++ src/Http/routes.php
@@ -5,8 +5,8 @@ use LivewireFilemanager\Filemanager\Http\Controllers\Api\BulkUploadController;
 use LivewireFilemanager\Filemanager\Http\Controllers\Api\FileController;
 use LivewireFilemanager\Filemanager\Http\Controllers\Api\FolderController;
 
-Route::prefix(config('livewire-fileuploader.api.prefix', 'filemanager/v1'))
-    ->middleware(config('livewire-fileuploader.api.middleware', ['api']))
+Route::prefix(config('livewire-filemanager.api.prefix', 'filemanager/v1'))
+    ->middleware(config('livewire-filemanager.api.middleware', ['api']))
     ->name('filemanager.api.')
     ->group(function () {
         Route::apiResource('folders', FolderController::class);