+++ src/Models/Folder.php
@@ -36,7 +36,7 @@ class Folder extends Model implements HasMedia
         });
 
         static::creating(function ($folder) {
-            if (! config('livewire-fileuploader.acl_enabled')) {
+            if (! config('livewire-filemanager.acl_enabled')) {
                 return;
             }
 
@@ -51,7 +51,7 @@ class Folder extends Model implements HasMedia
     protected static function booted()
     {
         static::addGlobalScope('user_id', function (Builder $builder) {
-            if (! config('livewire-fileuploader.acl_enabled')) {
+            if (! config('livewire-filemanager.acl_enabled')) {
                 return;
             }