+++ src/Traits/HasMediaOwner.php
@@ -9,7 +9,7 @@ trait HasMediaOwner
     protected static function booted()
     {
         static::addGlobalScope('user_id', function (Builder $builder) {
-            if (! config('livewire-fileuploader.acl_enabled')) {
+            if (! config('livewire-filemanager.acl_enabled')) {
                 return;
             }