+++ tests/Models/TestFolderModel.php
@@ -33,7 +33,7 @@ class TestFolderModel extends Model
         });
 
         static::creating(function ($folder) {
-            if (! config('livewire-fileuploader.acl_enabled')) {
+            if (! config('livewire-filemanager.acl_enabled')) {
                 return;
             }
 
@@ -48,7 +48,7 @@ class TestFolderModel extends Model
     protected static function booted()
     {
         static::addGlobalScope('user_id', function (Builder $builder) {
-            if (! config('livewire-fileuploader.acl_enabled')) {
+            if (! config('livewire-filemanager.acl_enabled')) {
                 return;
             }